backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-q2==1.7.3
psycopg2-binary==2.9.9
whitenoise==6.7.0
gunicorn==22.0.0
python-dotenv==1.0.1
